+++ src/app_data/container_state.rs
@@ -370,17 +370,19 @@ pub struct ContainerItem {
     pub state: State,
     pub status: String,
     pub tx: ByteStats,
+    pub is_oxker: bool,
 }
 
 impl ContainerItem {
     /// Create a new container item
     pub fn new(
-        id: ContainerId,
-        status: String,
-        image: String,
-        state: State,
-        name: String,
         created: u64,
+        id: ContainerId,
+        image: String,
+        is_oxker: bool,
+        name: String,
+        state: State,
+        status: String,
     ) -> Self {
         let mut docker_controls = StatefulList::new(DockerControls::gen_vec(state));
         docker_controls.start();
@@ -392,6 +394,7 @@ impl ContainerItem {
             docker_controls,
             id,
             image,
+            is_oxker,
             last_updated: 0,
             logs,
             mem_limit: ByteStats::default(),